[Refactoring] Organize classes in the correct modules
