+++ app/src/main/java/com/paulacr/blogviewer/feature/authors/AuthorsListViewModel.kt
@@ -6,7 +6,7 @@ import androidx.paging.PagingData
 import com.paulacr.blogviewer.BaseViewModel
 import com.paulacr.blogviewer.ViewState
 import com.paulacr.data.common.logError
-import com.paulacr.data.usecase.PostsUseCase
+import com.paulacr.domain.usecase.PostsUseCase
 import com.paulacr.domain.Author
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
