+++ app/src/main/java/com/paulacr/blogviewer/feature/posts/PostsViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import com.paulacr.blogviewer.BaseViewModel
 import com.paulacr.blogviewer.ViewState
 import com.paulacr.data.common.logError
-import com.paulacr.data.usecase.PostsUseCase
+import com.paulacr.domain.usecase.PostsUseCase
 import com.paulacr.domain.Post
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
