+++ app/src/main/res/layout/activity_authors_list.xml
@@ -5,16 +5,6 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <ProgressBar
-        android:id="@+id/loadingView"
-        android:visibility="visible"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"/>
-
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/authorsList"
         android:layout_width="match_parent"
@@ -48,5 +38,16 @@
         android:text="@string/errorLoadingAuthorsMessage"
         tools:visibility="visible"
         android:visibility="gone"/>
+
+    <ProgressBar
+        android:id="@+id/loadingView"
+        android:visibility="visible"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"/>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
