+++ app/src/main/res/values/dimens.xml
@@ -5,6 +5,7 @@
     <dimen name="default_dimen_16dp">16dp</dimen>
     <dimen name="default_dimen_4dp">4dp</dimen>
     <dimen name="default_dimen_8dp">8dp</dimen>
+    <dimen name="default_dimen_150dp">150dp</dimen>
 
     <!--Text Sizes-->
     <dimen name="default_text_size_16dp">16dp</dimen>
