+++ data/src/main/java/com/paulacr/data/mapper/AuthorMapper.kt
@@ -1,23 +1,19 @@
 package com.paulacr.data.mapper
 
 import com.paulacr.domain.Author
-import com.paulacr.domain.AuthorResponse
+import com.paulacr.data.responsedata.AuthorResponse
 import javax.inject.Inject
 
 class AuthorMapper @Inject constructor() {
 
-    fun map(rawData: List<AuthorResponse>): MutableList<Author> {
-        val listAuthors = mutableListOf<Author>()
+    fun map(rawData: List<AuthorResponse>): List<Author> =
 
-        rawData.forEach {
-            listAuthors.add(Author(
+        rawData.map {
+            Author(
                 id = it.id,
                 name = it.name,
                 username = it.username,
                 avatarUrl = it.avatarUrl
-            ))
+            )
         }
-
-        return listAuthors
-    }
 }
