+++ data/src/main/java/com/paulacr/data/repository/BlogRepository.kt
@@ -7,20 +7,13 @@ import androidx.paging.rxjava2.flowable
 import com.paulacr.domain.Author
 import com.paulacr.domain.Comment
 import com.paulacr.domain.Post
+import com.paulacr.domain.usecase.BlogRepository
 import io.reactivex.Flowable
 import io.reactivex.Single
 import javax.inject.Inject
 
-interface BlogRepository {
-
-    fun getAuthors(): Flowable<PagingData<Author>>
-
-    fun getPostsByAuthorId(id: Int): Single<List<Post>>
-
-    fun getCommentsByPostId(postId: Int): Single<List<Comment>>
-}
-
-class BlogRepositoryImpl @Inject constructor(private val pagingSource: GetAuthorsPagingSource) : BlogRepository {
+class BlogRepositoryImpl @Inject constructor(private val pagingSource: GetAuthorsPagingSource) :
+    BlogRepository {
 
     override fun getAuthors(): Flowable<PagingData<Author>> =
         Pager(
