+++ libDependencies.gradle
@@ -13,7 +13,7 @@ ext {
 
     // Network
     ext.retrofitVersion = '2.9.0'
-    ext.rxjavaVersion = '2.2.9'
+    ext.rxjavaVersion = '2.2.10'
     ext.RxJavaAdapterVersion = "2.9.0"
     ext.rxAndroidVersion = '2.1.1'
     ext.OkhttpVersion = '4.9.0'
